backend/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# ML dependencies
torch==2.1.2
torchvision==0.16.2
numpy==1.26.3
scipy==1.12.0
scikit-image==0.22.0
opencv-python-headless==4.9.0.80
Pillow==10.2.0

# Optional: Background removal (uncomment if needed)
# rembg==2.0.50

# Utilities
python-multipart==0.0.6

# Development
python-dotenv==1.0.0
